# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Webhook Pusher API",
    "version": "2.0.0",
    "description": "Webhook Pusher 是一个基于 EdgeOne 的消息推送服务，支持通过 Webhook 发送微信公众号消息。\n\n## 架构概述\n\n系统采用三层架构：\n- **Channel（渠道）**: 消息发送通道配置，如微信公众号\n- **App（应用）**: 业务应用，关联渠道并定义推送模式\n- **OpenID**: 应用下绑定的微信用户\n\n## 认证方式\n\n管理 API 需要在请求头中携带 Admin Token：\n```\nAuthorization: Bearer <admin_token>\n```\n或\n```\nX-Admin-Token: <admin_token>\n```\n\n## Webhook 调用\n\nWebhook API 无需认证，通过 App Key 标识应用：\n```\nGET /{appKey}.send?title=消息标题&desp=消息内容\nPOST /{appKey}.send\n```",
    "contact": {
      "name": "ixNieStudio",
      "email": "colin@ixnie.cn"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    }
  },
  "servers": [
    {
      "url": "/v1",
      "description": "API v1"
    }
  ],
...
  "paths": {
...
    "/config/": {
      "get": {
        "tags": [
          "Config"
        ],
        "summary": "获取系统配置",
        "operationId": "getConfig",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        },
        "description": "获取当前系统配置，敏感信息已脱敏",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "AdminToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "Config"
        ],
        "summary": "更新系统配置",
        "operationId": "putConfig",
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SystemConfig"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未授权"
          }
        },
        "description": "更新系统配置，如速率限制、数据保留策略等",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "AdminToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/config/reset-token": {
      "post": {
        "tags": [
          "Config"
        ],
        "summary": "重置管理员令牌",
        "operationId": "postConfigReset-token",
        "responses": {
          "200": {
            "description": "- 重置成功"
          },
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 0
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/ResetTokenResult"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "- 密码验证失败"
          },
          "401": {
            "description": "未授权"
          }
        },
        "description": "**自定义密码要求：** - 长度至少 12 个字符 - 包含至少一个大写字母 (A-Z) - 包含至少一个小写字母 (a-z) - 包含至少一个数字 (0-9) - 包含至少一个特殊字符 (!@#$%^&*()_+-=[]{}|;:,.<>?) **使用方式：** - 不提供参数：自动生成随机令牌（向后兼容） - 提供 newPassword 和 confirmPassword：使用自定义密码",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "AdminToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetTokenRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SystemConfig": {
        "type": "object",
        "properties": {
          "adminToken": {
            "type": "string"
          },
          "rateLimit": {
            "type": "object",
            "properties": {
              "perMinute": {
                "type": "number"
              }
            },
            "required": [
              "perMinute"
            ],
            "additionalProperties": false
          },
          "retention": {
            "type": "object",
            "properties": {
              "days": {
                "type": "number"
              }
            },
            "required": [
              "days"
            ],
            "additionalProperties": false
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "adminToken",
          "createdAt",
          "updatedAt"
        ],
        "description": "System 系统配置相关类型定义",
        "additionalProperties": false
      },
...
      "ResetTokenRequest": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "description": "自定义密码（可选）。需符合复杂度要求：至少12个字符，包含大小写字母、数字和特殊字符",
            "example": "MySecurePass123!"
          },
          "confirmPassword": {
            "type": "string",
            "description": "确认密码（使用自定义密码时必填）",
            "example": "MySecurePass123!"
          }
        }
      },
      "ResetTokenResult": {
        "type": "object",
        "properties": {
          "adminToken": {
            "type": "string",
            "description": "新的管理员令牌",
            "example": "AT_MySecurePass123!"
          },
          "message": {
            "type": "string",
            "description": "提示信息",
            "example": "管理员密码已重置为自定义密码。请妥善保管新密码，旧令牌已失效。"
          },
          "isCustomPassword": {
            "type": "boolean",
            "description": "是否使用自定义密码",
            "example": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin Token 认证"
      },
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token",
        "description": "Admin Token (alternative)"
      }
    }
  }
}